Make compaction framework format-insensitive
